商品アップデートと削除機能の修正 #12

+++ 230131_nestjs_study02/src/items/items.service.ts
@@ -1,4 +1,4 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import { BadRequestException, Injectable, NotFoundException } from '@nestjs/common';
 import { CreatItemDto } from './dto/creat-item.dto';
 import { ItemStatus } from './item-status.enum';
 import { Item } from '../entities/item.entity';
@@ -27,8 +27,11 @@ export class ItemsService {
         return this.itemRepository.creatItem(creatItemDto, user);
     }
 
-    async updateStatus(id: string): Promise<Item> {
+    async updateStatus(id: string, user: User): Promise<Item> {
         const item = await this.findById(id);
+        if(item.userId === user.id) {
+            throw new BadRequestException('この商品を購入することはできません');
+        }
         item.status = 
             item.status === ItemStatus.ON_SALE ?
             ItemStatus.SOLD_OUT :
@@ -38,7 +41,11 @@ export class ItemsService {
         return item;
     }
 
-    async delete(id: string): Promise<void> {
+    async delete(id: string, user: User): Promise<void> {
+        const item = await this.findById(id);
+        if(item.userId !== user.id) {
+            throw new BadRequestException('この商品を削除することはできません');
+        }
         await this.itemRepository.delete({ id });
     }
 }
